trees

+++ src/hog/hog.h
@@ -12,14 +12,14 @@ using std::vector;
 class Hog {
     int n_pixels_per_cel, n_cells_per_block;
     int n_bins; // 20 degrees
-    vector<float> getCellHistogram(const vector<vector<float>> &angles, const vector<vector<float>> &magn, const int &fromX, const int &fromY);
-    vector<float> l2blockNormalization(const vector<vector<float>>& cellsHistogram);
-    void computeGradient(const vector<vector<float>> &image, const int &posx, const int &posy, float &gradx, float &grady, const int &n_rows, const int &n_cols);
+    vector<double> getCellHistogram(const vector<vector<double>> &angles, const vector<vector<double>> &magn, const int &fromX, const int &fromY);
+    vector<double> l2blockNormalization(const vector<vector<double>>& cellsHistogram);
+    void computeGradient(const vector<vector<double>> &image, const int &posx, const int &posy, double &gradx, double &grady, const int &n_rows, const int &n_cols);
 public:
     Hog(int n_pixels_per_cel = 8, int n_cells_per_block = 1, int n_bins = 9);
 
-    vector<vector<float>> getHistogram(const vector<vector<float>> &image);
-    vector<float> getFlatHistogram(const vector<vector<float>> &image);
+    vector<vector<double>> getHistogram(const vector<vector<double>> &image);
+    vector<double> getFlatHistogram(const vector<vector<double>> &image);
 
 };
